+++ frontend/app/(tabs)/(audio)/_layout.jsx
@@ -15,15 +15,6 @@ const AudioLayout = () => {
         }}
       >
         <Stack.Screen name="index" options={{headerShown: false}} />
-        <Stack.Screen name="unified" options={{headerShown: false}} />
-        <Stack.Screen 
-          name="player" 
-          options={{
-            headerShown: false,
-            animation: 'slide_from_bottom',
-            animationDuration: 200,
-          }} 
-        />
       </Stack>
     </>
   );